Moving path related functions to own Tests class; adding tests for expected functionality
