+++ Coveralls.Tests/Coveralls.Tests.csproj
@@ -54,6 +54,7 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="PathToolsTests.cs" />
     <Compile Include="CoverageFileTests.cs" />
     <Compile Include="CoverallsBootstrapTests.cs" />
     <Compile Include="StringExtensionTests.cs" />
